src/detection_service/requirements.txt
fastapi
uvicorn
pydantic>=2.0
redis
python-dotenv
requests
